Fix warnings

+++ src/Helpers.cc
@@ -30,12 +30,14 @@ uint64_t ignition::math::Pair(const uint32_t _a, const uint32_t _b)
 /////////////////////////////////////////////
 std::tuple<uint32_t, uint32_t> ignition::math::Unpair(const uint64_t _key)
 {
-  // Must explicityly cast so that the _key is not auto cast to a double
+  // Must explicitly cast so that the _key is not auto cast to a double
   uint64_t sqrt = static_cast<uint64_t>(
       std::floor(std::sqrt(static_cast<long double>(_key))));
   uint64_t sq = sqrt * sqrt;
 
   return ((_key - sq) >= sqrt) ?
-    std::make_tuple(sqrt, _key - sq - sqrt) :
-    std::make_tuple(_key - sq, sqrt);
+    std::make_tuple(static_cast<uint32_t>(sqrt),
+                    static_cast<uint32_t>(_key - sq - sqrt)) :
+    std::make_tuple(static_cast<uint32_t>(_key - sq),
+                    static_cast<uint32_t>(sqrt));
 }
